+++ wasm/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0-dev
+
+- Provide a new way for creating a WasmModule via 
+  `wasmModuleCompileAsync` and `wasmModuleCompileSync`
+
 ## 0.2.0-dev
 
 - Update to Wasmer 2.1.0